Added DFLOAT compilation to all LDL. It compiles and runs but linalg tests fail at the moment

+++ lin_sys/direct/external/suitesparse/amd/src/amd_1.c
@@ -37,8 +37,8 @@ GLOBAL void AMD_1
     Int slen,		/* slen >= sum (Len [0..n-1]) + 7n,
 			 * ideally slen = 1.2 * sum (Len) + 8n */
     Int S [ ],		/* size slen workspace */
-    double Control [ ],	/* input array of size AMD_CONTROL */
-    double Info [ ]	/* output array of size AMD_INFO */
+    c_float Control [ ],	/* input array of size AMD_CONTROL */
+    c_float Info [ ]	/* output array of size AMD_INFO */
 )
 {
     Int i, j, k, p, pfree, iwlen, pj, p1, p2, pj2, *Iw, *Pe, *Nv, *Head,
